You can now click on different users.
Js: 
It now add a listener when you create a new contact. The listener checks what contact you clicked on, and then loads the chat window for that contact.
WIP: store the chat data when switching contact.

HTML:
Changed some classes, and deleted some as well.
Moved the javascript file outside of the body, so that it does not load before the DOM is loaded.

CSS:
Fixed some padding and margin issues.
Also fixed an issue where chatWindow outgrew itself, when using firefox.