Created a modal that list contacts
Js:
Created an array called contactList.
Created an getEventTarget function that returns the target that is clicked on.
Created a modal that is displayed only when the button is clicked and there is atleast one contact in contactList.
Added checks for white space in input fields.
Fixed the way the chat recognizes internett-picturefiles (.jpg .gif .png .svg)
Made a function to add contacts and eventlisteners to check if they are clicked.
Other functions made:
-removing child elements.
-enable button
-disable button

HTML:
Added a new button for choosing contacts.
Added a modal.

CSS:
Fixed some z-indexes and some max-width attributes.
Added css for the modal.
Added css for screens smaller than 750px